refactor: align 1cs intent naming and remove unused legacy swap machine
